requirements.txt
flask~=3.1.0
requests==2.26.0
mysql-connector-python~=9.2.0
python-telegram-bot~=21.10
selenium~=4.28.1
webdriver-manager~=4.0.2
bs4~=0.0.2
beautifulsoup4~=4.13.3
